+++ src/main/java/software/aoc/challenges/day01/PartA.java
@@ -4,7 +4,7 @@ import software.aoc.challenges.Solver;
 public final class PartA implements Solver {
 
     @Override
-    public int solve(String input) {
+    public long solve(String input) {
         return Dial.empty()
                 .follow(input)
                 .countTimesEndingAtZero();
